+++ CesiumAsync/include/CesiumAsync/Impl/unwrapFuture.h
@@ -20,8 +20,7 @@ auto futureFunctionToTaskFunction(Func&& f) {
     // Function taking no parameters
     if constexpr (isFuture<typename ContinuationReturnType<Func, void>::type>) {
       // And returning a Future
-      return
-          [f = std::forward<Func>(f)]() { return f()._task; };
+      return [f = std::forward<Func>(f)]() { return f()._task; };
     } else {
       // And returning a regular value
       return std::forward<Func>(f);